Transferred Character Select from MainActivity to CharacterSelectActivity
Cleaned up code
